feat(driver): Implement IO handling
Create functions to setup GPIO pins for the required configuration.
Craeted enum 1o_ports_e and assigned the pins to the named pins in 1o_e,
to provide the right pin and port mapping. Accessed the port names by
using a static arry in io.c as opposed to a switch statment.
Tested all pins as input and output in main.c.