feat(driver): Fixed issue with configuring all pins and tested the configuration with blinking the test led
IO pins PA13 and PA14 on the stm32l467rg are resevred mainly for SWD but
can be configured as limited GPIO pins. They are unused in the project,
so they were configured as default input, but that resulted in an error
during flashing. To fix the issue the pins were excluded from the IO
configuration. The configuration was then tested by blinking the
IO_TEST_LED.