feat(assert): Assert handler to make asserts in functions to alert for failures and Led driver
ASSERT define created using an assert handler that triggers when a
specific fault occurs, that is when  an assertion made is negated. The
handler works by first creating a debug BREAKPOINT in the function body,
then entering an endless loop, with TEST_LED blinking, until the
controller is reset. The assert handler is employed in the led driver to
ensure that TEST_LED has been configured properly, also to make sure the
configuration check only occurs once.
The led driver is not used in the assert handler to prevent recursive
assertions.