feat(assert): Static assert io_port_e
Added static assert to io.c to ensure that io_port_e enum is 1 byte.
Used a define marco check between the vscode intellisense and gcc
complier to remove errors in the editor. The error is because the vscode
intellisense recognizes enums as 4 bytes, so the static assertion of the
enum equalling 1 byte caused the error. Added complier flags
-fshort-enums to ensure gcc recognizes enums as 1 byte.