fix(test): Move test file to test.c and added test build to CI system
Moved all test functions to test.c and added the test to be used as a
complier function. That was done by adding the test name as a define in
the makefile. To avoid errors from the unused functions when a function
is tested the gcc unused attribute was used to ignore those functions.

A step to build all functions was added to the CI system. A bash script
to get all functions and build them is used to accomplish the test build
step in the CI system.