feat(driver): GPIO interrupt handling
Add interrupt handling for IO pins in io.c. IO interrupts are configured
by setting the necessary registers to setup the right EXTI lines,
interrupt masks and interrupt trigger. The interrupt service function is
then mapped to the right IO pin to complete the config. The interrupt is
then enabled in the enable function.
The EXTI handle functions for the 15 lines are then setup to check for
the right IO pin for the interrupt and execute the service function. The
interrupt handler functions are then tested to blink the TEST_LED.