fix(clock): Increased clock speed to 80Mhz
Increased clock speed from the default 4Mhz to the max value on the
microcontroller 80Mhz. To do this the PLL clock using the HSI clock as a
source, was set as the system clock.
The clock initilization was added to the mcu init function. The
BUSY_WAIT macro was also updated to reflect the 80Mhz clock speed.