feat(driver): UART driver and io fixes
Created a UART driver using a ring buffer to store string to be
transmitted. The ring buffer makes use of put and get functions to
update the buffer with checks for buffer fullness or emptiness before
adding or removing data. Interrupt handling for the UART trnamission was
employed, as polling is much more cpu exhaustive.
A simple temporary print function was used to test the UART driver. A
new assert macro was added for interrupts to be filled out later.
IO pins for UART TX and RX were updated and Alternate Function
intitialization for the IO pins was added to the io_init function.