feat(assert): Add TRACE to assert handler
Added tracing to assert handler, by making use of uart polling. The
TRCAE macro was not used to conserve flash space. The program counter
for the assertion is printed with the assert trace.
Addr2line is then used as a makefile rule to locate the file and line number
of the assertion.