feat(driver): IR Remote driver
Utilize NEC protocol to create a driver for the IR remote reception.
Used a timer interrupt to count the interval bettween pulses in the IR
message receieved by the IR_REMOTE pin interrupt. Made use of a ring
buffer to store the message command, exteracted from the 32 bit NEC
message.
Added tracing for the the received commands and tested the
functionality.