feat(pwm): PWM driver
PWM driver using the 2 channels of TIM3 to generate duty cycles based on
a percentage number passed. When a duty cycle of 0 is passed the
corresponsing channel is turned off and if at any time when a 0 value is
passed and both channels are subsequently turned off, TIM3 is disabled
to conserve power.