feat(driver): Drive Interface
Drive interface that interacts with the tb6612fng driver to simplify
application code. An array that maps a collection of unique directions
to appropriate motor speeds and modes ( e.g. CW, CCW) is utlized to set
the motor up to produce the correct motion.
Only basic diretcions,FORWARD and LEFT diretcions were included in the array, inorder to save
flash space, since the directions come in pairs e.g. (FORAWRD & REVRESE). The other directions were gotten by using a MODULO
operation in a preprocessor define, to match the direction to a basic
direction. The matched basic direction is then mirrored appropriately in
speed and/or direction.