feat(driver): ADC Driver
ADC driver for line sensor making use of continous ADC along wtih DMA
and an ADC interrupt. For the driver all 16 channels of ADC2 are read
for adc values, Direct Memory Access, DMA, is used to copy the adc
values from the ADC2 data reg to a location in memory. DMA is used in
the driver to reduce CPU loading during operation. A final memory
location is then used to store the adc values for use, by copying the
adc values from memory location 1 in an ADC interrupt to the final
location 2, this is done to avoid corruption of location 1 if it is read
from directly.
ADC functions to test a single ADC channel using polling is also
implemented in the driver for ADC testing. The ADC driver was tested by
reading 0 and 3.3V adc outputs when the corresponding adc pins were pulled
to those voltage levels.